# biddingApi/biddingApi/src/main/java/biddingApi/biddingApi/Model/BidDeleteResponse.java
package biddingApi.biddingApi.Model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
public @Data class BidDeleteResponse {

	private String status;

}